New request type FindItemLMT
This request tries to fetch all items that have been modified
after the specified timestamp

Such a query is required for implementing a bi-directional sync
with the server. However ASynK does not use this method, as such
a query does not give a complete picture of all the contacts on
the server and we are unable to accuratel distinguish between
contacts that have been deleted and modified on the server.